move ApiException
