+++ vip-api/src/main/java/fr/insalyon/creatis/vip/api/controller/AuthenticationController.java
@@ -1,11 +1,5 @@
 package fr.insalyon.creatis.vip.api.controller;
 
-import fr.insalyon.creatis.vip.api.business.ApiBusiness;
-import fr.insalyon.creatis.vip.api.business.ApiUserBusiness;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException;
-import fr.insalyon.creatis.vip.core.server.model.AuthenticationCredentials;
-import fr.insalyon.creatis.vip.core.server.model.AuthenticationInfo;
-import fr.insalyon.creatis.vip.api.model.ResetPasswordDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +8,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import fr.insalyon.creatis.vip.api.business.ApiBusiness;
+import fr.insalyon.creatis.vip.api.business.ApiUserBusiness;
+import fr.insalyon.creatis.vip.api.exception.ApiException;
+import fr.insalyon.creatis.vip.api.model.ResetPasswordDTO;
+import fr.insalyon.creatis.vip.core.server.model.AuthenticationCredentials;
+import fr.insalyon.creatis.vip.core.server.model.AuthenticationInfo;
 import jakarta.validation.Valid;
 
 @RestController
