+++ vip-api/src/main/java/fr/insalyon/creatis/vip/api/controller/RegisterUserController.java
@@ -1,18 +1,22 @@
 package fr.insalyon.creatis.vip.api.controller;
 
-import fr.insalyon.creatis.vip.api.business.ApiUserBusiness;
-import fr.insalyon.creatis.vip.api.model.SignUpUserDTO;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException;
-import fr.insalyon.creatis.vip.core.client.bean.User;
+import java.util.Date;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
+import fr.insalyon.creatis.vip.api.business.ApiUserBusiness;
+import fr.insalyon.creatis.vip.api.exception.ApiException;
+import fr.insalyon.creatis.vip.api.model.SignUpUserDTO;
+import fr.insalyon.creatis.vip.core.client.bean.User;
 import jakarta.validation.Valid;
-import java.util.Date;
 
 @RestController
 @RequestMapping("/register")
