+++ vip-api/src/main/java/fr/insalyon/creatis/vip/api/controller/data/DataController.java
@@ -1,11 +1,9 @@
 package fr.insalyon.creatis.vip.api.controller.data;
 
-import fr.insalyon.creatis.vip.api.business.DataApiBusiness;
-import fr.insalyon.creatis.vip.api.controller.ApiController;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException;
-import fr.insalyon.creatis.vip.api.model.ExistsApiResponse;
-import fr.insalyon.creatis.vip.api.model.PathProperties;
-import fr.insalyon.creatis.vip.api.model.UploadData;
+import java.io.File;
+import java.io.InputStream;
+import java.util.List;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,13 +12,20 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriUtils;
 
+import fr.insalyon.creatis.vip.api.business.DataApiBusiness;
+import fr.insalyon.creatis.vip.api.controller.ApiController;
+import fr.insalyon.creatis.vip.api.exception.ApiException;
+import fr.insalyon.creatis.vip.api.model.ExistsApiResponse;
+import fr.insalyon.creatis.vip.api.model.PathProperties;
+import fr.insalyon.creatis.vip.api.model.UploadData;
 import jakarta.servlet.http.HttpServletRequest;
-import java.io.File;
-import java.io.InputStream;
-import java.util.List;
 
 @RestController
 @RequestMapping("path")
