+++ vip-api/src/main/java/fr/insalyon/creatis/vip/api/controller/stats/StatsController.java
@@ -1,17 +1,18 @@
 package fr.insalyon.creatis.vip.api.controller.stats;
 
-import fr.insalyon.creatis.vip.api.business.StatsApiBusiness;
-import fr.insalyon.creatis.vip.api.controller.ApiController;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException.ApiError;
-import fr.insalyon.creatis.vip.api.model.stats.UsersList;
-import fr.insalyon.creatis.vip.api.model.stats.UsersNumber;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import fr.insalyon.creatis.vip.api.business.StatsApiBusiness;
+import fr.insalyon.creatis.vip.api.controller.ApiController;
+import fr.insalyon.creatis.vip.api.exception.ApiException;
+import fr.insalyon.creatis.vip.api.exception.ApiException.ApiError;
+import fr.insalyon.creatis.vip.api.model.stats.UsersList;
+import fr.insalyon.creatis.vip.api.model.stats.UsersNumber;
 
 @RestController
 @RequestMapping("/statistics")
