+++ vip-api/src/test/java/fr/insalyon/creatis/vip/api/rest/itest/SpringAuthenticationIT.java
@@ -11,7 +11,18 @@ import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequ
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.http.MediaType;
+import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors;
+
+import fr.insalyon.creatis.vip.api.exception.ApiException.ApiError;
+import fr.insalyon.creatis.vip.api.rest.config.BaseWebSpringIT;
+import fr.insalyon.creatis.vip.api.tools.spring.ApikeyRequestPostProcessor;
+import fr.insalyon.creatis.vip.api.tools.spring.BearerTokenRequestPostProcessor;
 
 /**
  * These tests check the authentication with the spring test tools.
