+++ vip-core/src/main/java/fr/insalyon/creatis/vip/core/server/security/common/RestExceptionHandler.java
@@ -1,9 +1,7 @@
-package fr.insalyon.creatis.vip.core.server.controller;
+package fr.insalyon.creatis.vip.core.server.security.common;
 
-import fr.insalyon.creatis.vip.core.server.exception.ApiException;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException.ApiError;
-import fr.insalyon.creatis.vip.core.server.model.ErrorCodeAndMessage;
-import fr.insalyon.creatis.vip.core.client.VipException;
+import java.util.Optional;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
@@ -20,7 +18,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import java.util.Optional;
+import fr.insalyon.creatis.vip.core.client.VipException;
+import fr.insalyon.creatis.vip.core.server.model.ErrorCodeAndMessage;
 
 @ControllerAdvice
 public class RestExceptionHandler extends ResponseEntityExceptionHandler {
