+++ vip-portal/src/test/java/fr/insalyon/creatis/vip/integrationtest/DefaultSpringConfigurationIT.java
@@ -31,17 +31,18 @@
  */
 package fr.insalyon.creatis.vip.integrationtest;
 
-import fr.insalyon.creatis.vip.api.controller.PlatformController;
-import fr.insalyon.creatis.vip.core.server.exception.ApiException;
-import fr.insalyon.creatis.vip.core.server.SpringCoreConfig;
-import fr.insalyon.creatis.vip.api.SpringRestApiConfig;
+import java.nio.file.Paths;
+
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
 import org.springframework.util.Assert;
 
-import java.nio.file.Paths;
+import fr.insalyon.creatis.vip.api.SpringRestApiConfig;
+import fr.insalyon.creatis.vip.api.controller.PlatformController;
+import fr.insalyon.creatis.vip.api.exception.ApiException;
+import fr.insalyon.creatis.vip.core.server.SpringCoreConfig;
 
 /**
  * Created by abonnet on 7/21/16.
